pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "red-team-arsenal"
version = "1.0.0"
description = "🔴💀 Ultimate Autonomous Red Team Arsenal — Rick Jefferson / RJ Business Solutions"
requires-python = ">=3.11"
dependencies = [
    # === Core AI ===
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "huggingface-hub>=0.24.0",
    "transformers>=4.44.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.3.0",
    # === Orchestration / Agents ===
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "crewai>=0.51.0",
    "pydantic>=2.8.0",
    # === Web / API ===
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "websockets>=13.0",
    "requests>=2.32.0",
    # === Research / Knowledge ===
    "tavily-python>=0.5.0",
    "apify-client>=1.7.0",
    "kaggle>=1.6.0",
    "scholarly>=1.7.0",
    "feedparser>=6.0.11",
    # === Data Processing ===
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "chromadb>=0.5.0",
    "faiss-cpu>=1.8.0",
    # === Cloudflare ===
    "cloudflare>=3.1.0",
    "boto3>=1.35.0",   # for R2 (S3-compatible)
    # === Security Tools (Python-native) ===
    "pyrit>=0.5.0",
    "agentic-radar>=0.2.0",
    # === Utilities ===
    "python-dotenv>=1.0.0",
    "rich>=13.8.0",
    "typer>=0.12.0",
    "loguru>=0.7.2",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    "markdown>=3.7",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
redteam = "main:app"
